server/requirements.txt




# --- Machine Learning & AI Frameworks ---
torch==2.8.0
torchvision==0.23.0
# REMOVED: triton (Primarily Linux-only)
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
faiss-cpu==1.12.0
scikit-learn==1.7.2

# --- Web Framework (FastAPI) & Server ---
fastapi==0.116.2
uvicorn==0.36.0
anyio==4.10.0
click==8.3.0
h11==0.16.0
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0

# --- Database & Backend Utilities ---
passlib
sqlalchemy
bcrypt
psycopg2-binary

# --- Data Science & Numerics ---
numpy==2.2.6
scipy==1.16.2
joblib==1.5.2
networkx==3.5
sympy==1.14.0
tqdm==4.67.1

# --- Core Utilities & Dependencies ---
annotated-types==0.7.0
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
regex==2025.9.18
setuptools==80.9.0
svgwrite==1.4.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
wcwidth==0.2.13
fastapi
uvicorn
python-dotenv
google-generativeai
requests
pillow
email-validator==2.1.1
dnspython==2.6.1
gunicorn